Updated Ethics Meter and Implemented
Did a few changes in scenes to test and deleted the old ethics script and replaced it with the new one. Got it working on game now, so is implemented. It just needs to be used in any situation that may affect the ethics. All that is left is to fix one bug and add comments to the code.